+++ lib/gpui/schema/component.ex
@@ -2,12 +2,13 @@ defmodule GPUI.Schema.Component do
   @moduledoc false
 
   @enforce_keys [:tag, :kind]
-  defstruct [:tag, :kind, events: [], attrs: []]
+  defstruct [:tag, :kind, events: [], attrs: [], children: false]
 
   @type t :: %__MODULE__{
           tag: atom(),
           kind: atom(),
           events: keyword(atom()),
-          attrs: keyword(atom())
+          attrs: keyword(atom()),
+          children: boolean()
         }
 end