+++ test/gpui/schema_test.exs
@@ -9,6 +9,7 @@ defmodule GPUI.SchemaTest do
              :ui_checkbox,
              :ui_input,
              :ui_select,
+             :ui_combobox,
              :span,
              :scroll,
              :list,
@@ -20,6 +21,7 @@ defmodule GPUI.SchemaTest do
            ]
 
     assert :"phx-change" in GPUI.Schema.events()
+    assert :"phx-search" in GPUI.Schema.events()
     assert :ui_select in GPUI.Schema.identified_tags()
     assert :raster in GPUI.Schema.resources()
     assert :border_radius in GPUI.Schema.styles()